Backend/requirements.txt
appnope==0.1.0
astroid==2.1.0
azure-cognitiveservices-speech==1.3.1
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.3.2
dill==0.2.9
Django==2.1.7
djangorestframework==3.9.2
en-core-web-sm==2.0.0
ffprobe==0.5
idna==2.8
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
numpy==1.16.2
pandas==0.24.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
pydub==0.23.1
Pygments==2.3.1
pylint==2.2.2
python-dateutil==2.8.0
pytz==2018.9
regex==2018.1.10
requests==2.21.0
six==1.12.0
spacy==2.0.18
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
traitlets==4.3.2
ujson==1.35
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
wrapt==1.10.11
